close files immediately after extracting
Prevents `open some/path/to/some/file: too many open files` from
happening.

The `defer` statement in Go executes deferred calls when the
function in which it was used returns. In this case, since all
files in the archive were being extracted in a loop inside a
single function, it kept handles for all archived files open
until the whole archive got extracted.